Move extensions to separate files
